+++ espr/src/ast/mod.rs
@@ -1,3 +1,4 @@
 //! Abstract Syntax Tree (AST) of EXPRESS Language
 
+pub mod expression;
 pub mod types;
