+++ espr/src/parser/entity/derive.rs
@@ -1,6 +1,6 @@
 use super::attribute::*;
 use crate::{
-    ast::types::*,
+    ast::{expression::*, types::*},
     parser::{combinator::*, expression::*, types::*},
 };
