+++ espr/src/parser/expression/simple.rs
@@ -1,139 +1,10 @@
 use super::{
-    super::{combinator::*, identifier::*, literal::*},
+    super::{combinator::*, identifier::*},
     aggregate_initializer::*,
     operator::*,
     primary::*,
 };
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum Expression {
-    Unary {
-        op: UnaryOperator,
-        arg: Box<Expression>,
-    },
-    Binary {
-        op: BinaryOperator,
-        arg1: Box<Expression>,
-        arg2: Box<Expression>,
-    },
-    Relation {
-        op: RelationOperator,
-        lhs: Box<Expression>,
-        rhs: Box<Expression>,
-    },
-    Literal(Literal),
-    QualifiableFactor {
-        factor: QualifiableFactor,
-        qualifiers: Vec<Qualifier>,
-    },
-    EntityConstructor {
-        name: String,
-        values: Vec<Expression>,
-    },
-    Interval {
-        op_low: IntervalOperator,
-        op_high: IntervalOperator,
-        high: Box<Expression>,
-        low: Box<Expression>,
-        item: Box<Expression>,
-    },
-    EnumerationReference {
-        ty: Option<String>,
-        enum_ref: String,
-    },
-    AggregateInitializer {
-        elements: Vec<Element>,
-    },
-    Query {
-        variable: String,
-        source: Box<Expression>,
-        expr: Box<Expression>,
-    },
-}
-
-macro_rules! impl_relation_op_expression {
-    ($f:ident, $op:path) => {
-        pub fn $f(self, other: Self) -> Self {
-            Expression::Relation {
-                op: $op,
-                lhs: Box::new(self),
-                rhs: Box::new(other),
-            }
-        }
-    };
-}
-
-macro_rules! impl_binary_op_expression {
-    ($f:ident, $op:path) => {
-        pub fn $f(self, other: Self) -> Self {
-            Expression::Binary {
-                op: $op,
-                arg1: Box::new(self),
-                arg2: Box::new(other),
-            }
-        }
-    };
-}
-
-impl Expression {
-    /// `SELF` constant
-    pub fn self_() -> Self {
-        Self::self_qualified(Vec::new())
-    }
-
-    /// `SELF` constant with qualifiers
-    pub fn self_qualified(qualifiers: Vec<Qualifier>) -> Self {
-        Expression::QualifiableFactor {
-            factor: QualifiableFactor::BuiltInConstant(BuiltInConstant::SELF),
-            qualifiers,
-        }
-    }
-
-    /// `?` constant
-    pub fn indeterminate() -> Self {
-        Expression::QualifiableFactor {
-            factor: QualifiableFactor::BuiltInConstant(BuiltInConstant::INDETERMINATE),
-            qualifiers: Vec::new(),
-        }
-    }
-
-    /// Real value literal
-    pub fn real(value: f64) -> Self {
-        Expression::Literal(Literal::Real(value))
-    }
-
-    impl_relation_op_expression!(leq, RelationOperator::LEQ);
-    impl_relation_op_expression!(geq, RelationOperator::GEQ);
-    impl_relation_op_expression!(lt, RelationOperator::LT);
-    impl_relation_op_expression!(gt, RelationOperator::GT);
-    impl_relation_op_expression!(eq, RelationOperator::Equal);
-    impl_relation_op_expression!(neq, RelationOperator::NotEqual);
-    impl_relation_op_expression!(in_, RelationOperator::In);
-    impl_relation_op_expression!(like, RelationOperator::Like);
-
-    impl_binary_op_expression!(and, BinaryOperator::And);
-    impl_binary_op_expression!(or, BinaryOperator::Or);
-    impl_binary_op_expression!(xor, BinaryOperator::Xor);
-}
-
-macro_rules! impl_binary_op_expression {
-    ($f:ident, $ops:path, $op:path) => {
-        impl $ops for Expression {
-            type Output = Self;
-            fn $f(self, other: Expression) -> Self {
-                Expression::Binary {
-                    op: $op,
-                    arg1: Box::new(self),
-                    arg2: Box::new(other),
-                }
-            }
-        }
-    };
-}
-
-impl_binary_op_expression!(add, std::ops::Add, BinaryOperator::Add);
-impl_binary_op_expression!(sub, std::ops::Sub, BinaryOperator::Sub);
-impl_binary_op_expression!(mul, std::ops::Mul, BinaryOperator::Mul);
+use crate::ast::expression::*;
 
 /// Create left-joined tree `1.0 + 2.0 - 3.0` into `(- (+ 1.0 2.0) 3.0)`
 fn create_tree(mut head: Expression, tails: Vec<(BinaryOperator, Expression)>) -> Expression {
