+++ espr/src/parser/types/concrete.rs
@@ -2,6 +2,7 @@ use super::{
     super::{combinator::*, expression::*, identifier::*},
     *,
 };
+use crate::ast::expression::*;
 
 /// 193 concrete_types = [aggregation_types] | [simple_types] | [type_ref].
 pub fn concrete_types(input: &str) -> ParseResult<ConcreteType> {
